NFC: Move shared validation helpers into HLSLTestDataTypes.h
Move isFloatingPointType, ValidationType, ValidationConfig,
DefaultValidation, StrictValidation, and doValuesMatch overloads
from LongVectors.cpp and LinearAlgebra.cpp into the shared header.
Both files now use 'using' declarations to reference them.
